requirements.txt
Flask
flask-cors
python-dotenv
langchain
langgraph
langchain-google-genai
google-generativeai
python-docx
google-cloud-aiplatform
